+++ lib/online_migrations/background_migrations/migration.rb
@@ -138,7 +138,7 @@ module OnlineMigrations
             progresses.sum.round(2)
           end
         elsif rows_count
-          if rows_count > 0
+          if rows_count > 0 && rows_count > batch_size
             jobs_rows_count = migration_jobs.succeeded.sum(:batch_size)
             # The last migration job may need to process the amount of rows
             # less than the batch size, so we can get a value > 1.0.
@@ -183,14 +183,14 @@ module OnlineMigrations
       def retry
         if composite? && failed?
           children.failed.each(&:retry)
-          running!
+          enqueued!
           true
         elsif failed?
           iterator = BatchIterator.new(migration_jobs.failed)
           iterator.each_batch(of: 100) do |batch|
             batch.each(&:retry)
           end
-          running!
+          enqueued!
           true
         else
           false